refactor: extract skills and session helpers
