+++ tests/unit/session-query.test.mjs
@@ -9,8 +9,8 @@ const __dirname = path.dirname(__filename);
 const require = createRequire(import.meta.url);
 const fs = require('fs');
 
-const cliPath = path.join(__dirname, '..', '..', 'cli.js');
-const cliContent = fs.readFileSync(cliPath, 'utf-8');
+const sessionsPath = path.join(__dirname, '..', '..', 'lib', 'cli-sessions.js');
+const cliContent = fs.readFileSync(sessionsPath, 'utf-8');
 
 // Extract the functions we need to test
 function extractFunction(content, funcName) {
